feat: integer world-market prices and catalog fallback (Refs #3093)
Slice 1 of #3093 — foundational price model and display:

- WorldMarketState.prices is Map<CommodityId, int>; fromJson floors legacy
  doubles and clamps negatives to 0 for backward-compatible saves.
- world_market_phase floors PriceDiscovery results at persistence boundary;
  DealMatcher still receives double prices at the call site.
- ResourceRules.defaultMarketPriceForCommodityId exposes catalog defaults
  for raw-resource commodities.
- Trade screen Market tab renders integer prices with catalog fallback
  for unseeded raw resources; manufactured commodities still show em-dash
  until in-game price discovery (follow-up slice).
- SPEC/game/world-market.md and SPEC/ui/trade-screen.md updated.
- Tests across models, data, logic, AI, and app packages.

+++ app/lib/widgetbook/catalog_part6.dart
@@ -291,12 +291,14 @@ Game _tradeScreenStoryGame() {
   // aggregate volumes for a handful of commodities — the remaining
   // rows render the em-dash price glyph + `Bids 0 / Offers 0` zero
   // default so reviewers can see both code paths at a glance.
-  const Map<CommodityId, double> prices = <CommodityId, double>{
-    'timber': 30.0,
-    'iron': 80.0,
-    'grain': 50.0,
-    'fabric': 120.0,
-    'castIron': 175.0,
+  // Post-#3093: WorldMarketState.prices is `Map<CommodityId, int>` (floored at
+  // persistence boundary per SPEC/game/world-market.md § Price discovery).
+  const Map<CommodityId, int> prices = <CommodityId, int>{
+    'timber': 30,
+    'iron': 80,
+    'grain': 50,
+    'fabric': 120,
+    'castIron': 175,
   };
   const Map<CommodityId, MarketActivity> activity =
       <CommodityId, MarketActivity>{
